third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/fetch-event-referrer-policy.https-expected.txt

This is a testharness.js-based test.
FAIL Service Worker responds to fetch event with the referrer policy assert_unreached: unexpected rejection: assert_equals: Service Worker should respond to fetch with the referrer URL when a member of RequestInit is present expected "Referrer: https://web-platform.test:8444/service-workers/service-worker/fake-referrer\nReferrerPolicy: no-referrer-when-downgrade" but got "Referrer: https://web-platform.test:8444/service-workers/service-worker/resources/fake-referrer\nReferrerPolicy: no-referrer-when-downgrade" Reached unreachable code
Harness: the test ran to completion.
